python_service/requirements.txt
# Phase 4: Python Service Dependencies

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.9

# Environment Variables
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# OCR (for scanned PDFs) - Optional for now
# pytesseract==0.3.10
# pdf2image==1.16.3

# Image Processing (newer version that installs easier)
Pillow>=10.0.0

# Machine Learning & Embeddings (Phase 6) - Python 3.13 compatible
transformers==4.36.0
torch==2.1.2

# Vector Database (Phase 7)
faiss-cpu==1.7.4
numpy==1.24.3

# OpenAI Integration (Phase 8)
openai==1.3.0

# HTTP Client
httpx==0.25.0
requests==2.31.0

# Data Processing
pandas==2.1.3

# Logging
loguru==0.7.2
